Introduce concept of optional characters, for example space can be matched by either a another space or a slash. Non optional characters are called core.
Strict isMatch is applied only on core characters of query. Soft score have access to the full query for scoring. This replace the regex that strip space from query and increase precision for multi word query eg "git push".

To be able to do that without recomputing core for each item, it is done at the filter() level.