Allow to interchange forward and backward slashes in query.
Not full support for interchange because we still use path.sep to count directory depth and so on.

However this extend the "space match path sep" to support wrong slash as well as colon. This allow translation of php/rail namespace to folder path. This could also be used for a file used in a multi OS environment.